requirements.txt
flair==0.10
numpy==1.18.1
pandas==1.0.1
python-dotenv==0.20.0
requests==2.24.0
spacy==3.2.3
SPARQLWrapper==2.0.0
